server/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

beautifulsoup4==4.12.3
certifi==2024.7.4
duckdb==1.0.0
eld==1.0.6
fastapi==0.111.1
httpx==0.27.0
ingredient_parser_nlp==1.0.0
nltk==3.8.1
pandas==2.2.2
pydantic==2.8.2
pymongo==4.8.0
python-dotenv==1.0.1
recipe_scrapers==14.58.0
spacy==3.7.5
starlette==0.37.2
transformers==4.42.4

aiohttp~=3.9.5
cachetools~=5.4.0
ingredient_parser~=1.0.2
